Keep filtering code and adjudication code in separate files
